Experiment with a signature parser that borrows from input &str
To avoid having JavaType::Object need to own a String for the object
name, this experiments with being able to parse signatures that borrow
names from the input string to reduce the number of allocations
required.

Although this does update the parser itself, it's awkward to then deal
with TypeSignature in other places if it needs a lifetime (e.g. can't be
included in Errors)