CI: Enable Android builds
This is also a general refresh of ci.yml which hasn't been touch in a while

This adds steps to build docs with `-Dwarning` and fixes the
various warnings that were found with that.

The invocation feature has been updated so that if it's enabled on
Android then any attempt to use `JavaVM::new` or `JavaVM::with_libjvm`
will simply return an `Unsupported` error.

Criterion was updated to 0.6

libloading and java-locator are not added as dependencies on Android,
when building with the invocation feature.

There is now an MSRV build that checks building with the rust-version
we specify in Cargo.toml (to check that it's still accurate).

The bare-minimum version for java-locator was tweaked to 0.1.3 to avoid
MSRV build failures from depending on glob 0.2